Add Windows testing support on CI (#16)
* Add Windows testing support on CI

I recently added tests for the Windows platform on a tool using this library and it failed while running this libraries manifest file. Thus I believe it's probably a good idea to test for Windows support right within this repo.

The exact command that hung up for 6 hours was:
```
C:\Users\runneradmin\AppData\Local\Temp\TemporaryDirectory.vFcVbW\swift-custom-dump-manifest.exe -handle 674
```

* Limit Windows testing to 30 minutes

As the issue is a specific command hanging forever, it's a good idea to limit the windows tests so the CI doesn't unnecessarily block CI resources for 6 hours (GitHubs hard limit). #SaveTheEnvironment

* Update DumpTests.swift

Co-authored-by: Stephen Celis <stephen.celis@gmail.com>