Fix `CustomDumpRepresentable` max depth
When we reach a custom dump representable value, we should not decrement
the max depth when printing it.

At times this can cause the max depth to roll from 0 to -1, which
currently causes the logic that collapses nodes to leave them expanded.

We should update that logic, as well, so I'll do that in another commit.