Filter properties prefixed with `_$`
This seems to be the macro convention, and as folks add `@Observable` to
their types they will begin to see these details leaking into their
dumps, which is generally not relevant and doesn't result in a dump that
resembles the original source.

We could consider a configuration option in the future to dump these
values, but for now folks can always use plain `dump` or a mirror to
access these values when they need them.